Bunch of updates for c++11 - c++17.
Some scoped enums and Message/Event/TransactionId support.
NULL -> nullptr.
override on derived virtual functions.
Use of auto and unique_ptr in selected places.
Still much work to do.

+++ DcrLib/ScreenTable_t.h
@@ -39,7 +39,7 @@ struct ScreenTable_t
     size_t        OffsetY;
     size_t        ColumnCount;
 
-    const size_t* pColumnWidths;
+    const int*    pColumnWidths;
     const RECT*   pTextRects;
 };
